# client/src/components/swagger.yaml
openapi: 3.0.0
info:
  title: Phong Tro API
  description: API quản lý phòng trọ - Room Rental Management System
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: 'http://localhost:3000/api'
    description: Development Server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  fullName:
                    type: string
                  token:
                    type: string
                  refreshToken:
                    type: string
        '400':
          description: Invalid input data
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  token:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid refresh token
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  # ===== PROPERTIES ENDPOINTS =====
  /properties:
    post:
      tags:
        - Properties
      summary: Create new property
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyDto'
      responses:
        '201':
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '401':
          description: Unauthorized

    get:
      tags:
        - Properties
      summary: Get all properties
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
        '401':
          description: Unauthorized
  /properties/{id}:
    get:
      tags:
        - Properties
      summary: Get property by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Property not found

    put:
      tags:
        - Properties
      summary: Update property
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyDto'
      responses:
        '200':
          description: Property updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Property not found

    delete:
      tags:
        - Properties
      summary: Delete property
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Property deleted successfully
        '404':
          description: Property not found

  # ===== ROOMS ENDPOINTS =====
  /rooms:
    post:
      tags:
        - Rooms
      summary: Create a new room
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomDto'
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'

    get:
      tags:
        - Rooms
      summary: Get all rooms with filters
      security:
        - BearerAuth: []
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [VACANT, OCCUPIED, MAINTENANCE, RESERVED]
        - name: type
          in: query
          schema:
            type: string
            enum: [SINGLE, DOUBLE, DELUXE]
        - name: floor
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/stats:
    get:
      tags:
        - Rooms
      summary: Get room statistics
      security:
        - BearerAuth: []
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Room statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRooms:
                    type: number
                  vacantRooms:
                    type: number
                  occupiedRooms:
                    type: number
                  maintenanceRooms:
                    type: number
                  occupancyRate:
                    type: number
  /rooms/{id}:
    get:
      tags:
        - Rooms
      summary: Get room by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found

    patch:
      tags:
        - Rooms
      summary: Update room
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomDto'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'

    delete:
      tags:
        - Rooms
      summary: Delete room
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room deleted successfully

  # ===== TENANTS ENDPOINTS =====
  /tenants:
    post:
      tags:
        - Tenants
      summary: Create new tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantDto'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'

    get:
      tags:
        - Tenants
      summary: Get all tenants
      security:
        - BearerAuth: []
      parameters:
        - name: propertyId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
  /tenants/{id}:
    get:
      tags:
        - Tenants
      summary: Get tenant by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          description: Tenant not found

    put:
      tags:
        - Tenants
      summary: Update tenant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantDto'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'

    delete:
      tags:
        - Tenants
      summary: Delete tenant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant deleted successfully

  # ===== INVOICES ENDPOINTS =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterDto:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [ADMIN, STAFF, MANAGER]
          default: STAFF
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    # Property DTOs
    CreatePropertyDto:
      type: object
      required:
        - name
        - address
        - totalFloors
        - totalRooms
      properties:
        name:
          type: string
        address:
          type: string
        totalFloors:
          type: number
        totalRooms:
          type: number
        currency:
          type: string
          default: VND
        timezone:
          type: string
          default: Asia/Ho_Chi_Minh
        settings:
          type: object
    UpdatePropertyDto:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        totalFloors:
          type: number
        totalRooms:
          type: number
        currency:
          type: string
        timezone:
          type: string
        settings:
          type: object
    Property:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        totalFloors:
          type: number
        totalRooms:
          type: number
        currency:
          type: string
        timezone:
          type: string
        ownerId:
          type: string
        settings:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Room DTOs
    CreateRoomDto:
      type: object
      required:
        - propertyId
        - roomNumber
        - floor
        - type
        - price
        - deposit
      properties:
        propertyId:
          type: string
        roomNumber:
          type: string
        floor:
          type: number
          minimum: 1
        type:
          type: string
          enum: [SINGLE, DOUBLE, DELUXE]
        price:
          type: number
          minimum: 0
        deposit:
          type: number
          minimum: 0
        area:
          type: number
          minimum: 0
        amenities:
          type: object
        description:
          type: string
        images:
          type: array
          items:
            type: string
    UpdateRoomDto:
      type: object
      properties:
        roomNumber:
          type: string
        floor:
          type: number
          minimum: 1
        type:
          type: string
          enum: [SINGLE, DOUBLE, DELUXE]
        status:
          type: string
          enum: [VACANT, OCCUPIED, MAINTENANCE, RESERVED]
        price:
          type: number
          minimum: 0
        deposit:
          type: number
          minimum: 0
        area:
          type: number
          minimum: 0
        amenities:
          type: object
        description:
          type: string
        images:
          type: array
          items:
            type: string
    Room:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        roomNumber:
          type: string
        floor:
          type: number
        type:
          type: string
        status:
          type: string
        price:
          type: number
        deposit:
          type: number
        area:
          type: number
        amenities:
          type: object
        description:
          type: string
        images:
          type: array
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Tenant DTOs
    CreateTenantDto:
      type: object
      required:
        - propertyId
        - fullName
        - phone
      properties:
        propertyId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        idCard:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        address:
          type: string
        emergencyContact:
          type: object
    UpdateTenantDto:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        idCard:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        address:
          type: string
        emergencyContact:
          type: object
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
    Tenant:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        idCard:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        address:
          type: string
        emergencyContact:
          type: object
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Invoice DTOs
...